bobob/requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==5.1.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
pytz==2024.2
PyYAML==6.0.2
setuptools==75.3.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
python-dotenv==1.0.1
